+++ src/MenuModal.tsx
@@ -3,14 +3,14 @@ export { MenuModal }
 import React from 'react'
 import { usePageContext } from './renderer/usePageContext'
 import { css } from './utils/css'
-import { viewDesktop, viewTablet } from './Layout'
+import { bodyMaxWidth, viewDesktop, viewTablet, scrollFadeMask } from './Layout'
 import { ExternalLinks } from './ExternalLinks'
 import { Style } from './utils/Style'
 import { NavigationWithColumnLayout } from './MenuModal/NavigationWithColumnLayout'
 import { closeMenuModal, closeMenuModalOnMouseLeave, keepMenuModalOpenOnMouseOver } from './MenuModal/toggleMenuModal'
 import { EditLink } from './EditLink'
 
-function MenuModal({ isTopNav }: { isTopNav: boolean }) {
+function MenuModal({ isTopNav, isNavLeftAlwaysHidden_ }: { isTopNav: boolean; isNavLeftAlwaysHidden_: boolean }) {
   return (
     <>
       <Style>{getStyle()}</Style>
@@ -21,11 +21,15 @@ function MenuModal({ isTopNav }: { isTopNav: boolean }) {
           position: isTopNav ? 'absolute' : 'fixed',
           width: '100%',
           top: 'var(--nav-head-height)',
-          left: 0,
           zIndex: 199, // maximum value, because docsearch's modal has `z-index: 200`
           background: '#ededef',
           transitionProperty: 'opacity',
           transitionTimingFunction: 'ease',
+          maxWidth: isNavLeftAlwaysHidden_ ? undefined : bodyMaxWidth,
+          // Horizontal align
+          // https://stackoverflow.com/questions/3157372/css-horizontal-centering-of-a-fixed-div/32694476#32694476
+          left: '50%',
+          transform: 'translateX(-50%)',
         }}
         onMouseOver={keepMenuModalOpenOnMouseOver}
         onMouseLeave={closeMenuModalOnMouseLeave}
@@ -37,6 +41,7 @@ function MenuModal({ isTopNav }: { isTopNav: boolean }) {
             overflowY: 'scroll',
             // We don't set `container` to the parent #menu-modal-wrapper beacuse of a Chrome bug (showing a blank <MenuModal>). Edit: IIRC because #menu-modal-wrapper has `position: fixed`.
             container: 'container-viewport / inline-size',
+            ...scrollFadeMask,
           }}
         >
           <Nav />
@@ -66,7 +71,7 @@ function BorderBottom() {
     <div
       id="border-bottom"
       style={{
-        background: '#fff',
+        background: 'var(--color-bg-white)',
         height: 'var(--block-margin)',
         width: '100%',
       }}