+++ src/vite.config.ts
@@ -46,5 +46,4 @@ const config: UserConfig = {
   // ▲ [WARNING] Transforming this CSS nesting syntax is not supported in the configured target environment ("chrome87", "edge88", "es2020", "firefox78", "safari14") [unsupported-css-nesting]
   // ```
   build: { target: 'es2022' },
-  clearScreen: false,
 }